added converstion from string format of date in Excel to Date format

+++ src/main/java/com/financus/realestates/RealEstatePeriods.java
@@ -5,9 +5,9 @@ import java.util.Date;
 public class RealEstatePeriods {
     Date date;
     String city;
-    int price;
+    double price;
 
-    public RealEstatePeriods(Date date, String city, int price) {
+    public RealEstatePeriods(Date date, String city, double price) {
         this.date = date;
         this.city = city;
         this.price = price;
